requirements.txt
flask==2.3.3
streamlit==1.28.0
pandas==2.1.1
numpy==1.26.0
matplotlib==3.8.0
seaborn==0.13.0
requests==2.31.0
python-dotenv==1.0.0
yfinance==0.2.31
openai==0.28.1
scikit-learn==1.3.1
newsapi-python==0.2.7
